% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_intersect.R
\name{all_intersect}
\alias{all_intersect}
\title{all_intersect}
\usage{
all_intersect(geom)
}
\arguments{
\item{geom}{()}
}
\value{
the geometry that contain the intersect of all geom contain in param
}
\description{
all_intersect
}
